Update schedules in select platform
