+++ homeassistant/components/netatmo/climate.py
@@ -58,6 +58,7 @@ from .data_handler import (
     HOMESTATUS_DATA_CLASS_NAME,
     NetatmoDataHandler,
 )
+from .helper import get_all_home_ids
 from .netatmo_entity_base import NetatmoBase
 
 _LOGGER = logging.getLogger(__name__)
@@ -131,9 +132,6 @@ async def async_setup_entry(
     if not home_data or home_data.raw_data == {}:
         raise PlatformNotReady
 
-    if HOMEDATA_DATA_CLASS_NAME not in data_handler.data:
-        raise PlatformNotReady
-
     entities = []
     for home_id in get_all_home_ids(home_data):
         for room_id in home_data.rooms[home_id]:
@@ -145,12 +143,18 @@ async def async_setup_entry(
             if home_status and room_id in home_status.rooms:
                 entities.append(NetatmoThermostat(data_handler, home_id, room_id))
 
-        hass.data[DOMAIN][DATA_SCHEDULES][home_id] = {
-            schedule_id: schedule_data.get("name")
-            for schedule_id, schedule_data in (
-                data_handler.data[HOMEDATA_DATA_CLASS_NAME].schedules[home_id].items()
-            )
-        }
+        hass.data[DOMAIN][DATA_SCHEDULES].update(
+            {
+                home_id: {
+                    schedule_id: schedule_data.get("name")
+                    for schedule_id, schedule_data in (
+                        data_handler.data[HOMEDATA_DATA_CLASS_NAME]
+                        .schedules[home_id]
+                        .items()
+                    )
+                }
+            }
+        )
 
     hass.data[DOMAIN][DATA_HOMES] = {
         home_id: home_data.get("name")
@@ -617,14 +621,3 @@ class NetatmoThermostat(NetatmoBase, ClimateEntity):
         device_info: DeviceInfo = super().device_info
         device_info["suggested_area"] = self._room_data["name"]
         return device_info
-
-
-def get_all_home_ids(home_data: pyatmo.HomeData | None) -> list[str]:
-    """Get all the home ids returned by NetAtmo API."""
-    if home_data is None:
-        return []
-    return [
-        home_data.homes[home_id]["id"]
-        for home_id in home_data.homes
-        if "modules" in home_data.homes[home_id]
-    ]
