+++ homeassistant/components/netatmo/helper.py
@@ -1,6 +1,10 @@
 """Helper for Netatmo integration."""
+from __future__ import annotations
+
 from dataclasses import dataclass
 from uuid import UUID, uuid4
+
+import pyatmo
 
 
 @dataclass
@@ -15,3 +19,14 @@ class NetatmoArea:
     mode: str
     show_on_map: bool
     uuid: UUID = uuid4()
+
+
+def get_all_home_ids(home_data: pyatmo.HomeData | None) -> list[str]:
+    """Get all the home ids returned by NetAtmo API."""
+    if home_data is None:
+        return []
+    return [
+        home_data.homes[home_id]["id"]
+        for home_id in home_data.homes
+        if "modules" in home_data.homes[home_id]
+    ]
